+++ src/utils/sleep.ts
@@ -1,0 +1,4 @@
+/** Resolve after `ms` milliseconds. */
+export function sleep(ms: number): Promise<void> {
+  return new Promise((resolve) => setTimeout(resolve, ms));
+}
